+++ src/lib.rs
@@ -1,3 +1,5 @@
+#![warn(unsafe_op_in_unsafe_fn)]
+
 pub mod cli;
 pub mod compiler;
 pub mod dotenv;