+++ src/vm/mod.rs
@@ -1,5 +1,6 @@
 pub mod gc;
 pub mod interpreter;
+#[cfg(feature = "jit")]
 pub mod jit;
 pub mod well_known;